Fix various display issues w/ new Checkbox
FinalCheckbox was changed to use the new Checkbox component in #6528,
but many places still passed `toggle` instead of the new `showAsToggle`
via the `FinalCheckbox` component to it.

Also changed some places to use the new component (there are more
though, but some of them require more style tweaks to not look weird).